sahur_claude/sahur/sahur-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sahur-mcp"
version = "0.1.0"
description = "MCP proxy server for SAHUR"
readme = "README.md"
authors = [
    {name = "SAHUR Team"}
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "mcp-server>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["sahur_mcp"]

[project.scripts]
sahur-mcp = "sahur_mcp.main:main"
